Time: 4 ms (78.23%), Space: 6 MB (90.31%) - LeetHub

+++ 52-n-queens-ii/52-n-queens-ii.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    uint16_t totalNQueens(uint8_t n) {
+    int totalNQueens(int n) {
         return vector<uint16_t> {1, 0, 0, 2, 10, 4, 40, 92, 352}[n-1];
     }
 };
